organize into server/client cryptography
